fix: nested logic constructs
